+++ src/DXRenderBufferPool.hpp
@@ -6,9 +6,9 @@ namespace BEbraEngine {
 	class DXRenderBufferPool : public IRenderBufferPool
 	{
 	public:
-		void allocate(size_t count, size_t sizeofData, AbstractRender::TypeBuffer type) override;
+		void allocate(uint32_t count, uint32_t sizeofData, AbstractRender::TypeBuffer type) override;
 		void free(std::weak_ptr<RenderBufferView> obj) override;
-		void deallocate(size_t count) override;
+		void deallocate(uint32_t count) override;
 		void setContext(AbstractRender* render) override;
 		std::shared_ptr<RenderBuffer> getBuffer() override;
 		size_t getCount() override;
